requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.35.0
httpx==0.27.0
requests==2.32.3
apscheduler==3.10.4
python-multipart==0.0.9
pytz==2024.1
openpyxl==3.1.4
beautifulsoup4==4.12.3
pdfplumber==0.11.1
